Invoker refactor 1
