+++ service/access/server_internal_test.go
@@ -1021,10 +1021,10 @@ func TestServer_ExecuteScriptAtBlockHeight(t *testing.T) {
 		t.Parallel()
 
 		invoker := mocks.BaselineInvoker(t)
-		invoker.ScriptFunc = func(height uint64, script []byte, parameters []cadence.Value) (cadence.Value, error) {
+		invoker.ScriptFunc = func(height uint64, script []byte, parameters [][]byte) (cadence.Value, error) {
 			assert.Equal(t, mocks.GenericHeight, height)
 			assert.Equal(t, mocks.GenericBytes, script)
-			assert.Equal(t, []cadence.Value{cadenceValue}, parameters)
+			assert.Equal(t, [][]byte{cadenceValueBytes}, parameters)
 
 			return mocks.GenericAmount(0), nil
 		}
@@ -1047,7 +1047,7 @@ func TestServer_ExecuteScriptAtBlockHeight(t *testing.T) {
 		t.Parallel()
 
 		invoker := mocks.BaselineInvoker(t)
-		invoker.ScriptFunc = func(uint64, []byte, []cadence.Value) (cadence.Value, error) {
+		invoker.ScriptFunc = func(uint64, []byte, [][]byte) (cadence.Value, error) {
 			return nil, mocks.GenericError
 		}
 
@@ -1079,10 +1079,10 @@ func TestServer_ExecuteScriptAtBlockID(t *testing.T) {
 		t.Parallel()
 
 		invoker := mocks.BaselineInvoker(t)
-		invoker.ScriptFunc = func(height uint64, script []byte, parameters []cadence.Value) (cadence.Value, error) {
+		invoker.ScriptFunc = func(height uint64, script []byte, parameters [][]byte) (cadence.Value, error) {
 			assert.Equal(t, mocks.GenericHeight, height)
 			assert.Equal(t, mocks.GenericBytes, script)
-			assert.Equal(t, []cadence.Value{cadenceValue}, parameters)
+			assert.Equal(t, [][]byte{cadenceValueBytes}, parameters)
 
 			return mocks.GenericAmount(0), nil
 		}
@@ -1135,7 +1135,7 @@ func TestServer_ExecuteScriptAtBlockID(t *testing.T) {
 		t.Parallel()
 
 		invoker := mocks.BaselineInvoker(t)
-		invoker.ScriptFunc = func(uint64, []byte, []cadence.Value) (cadence.Value, error) {
+		invoker.ScriptFunc = func(uint64, []byte, [][]byte) (cadence.Value, error) {
 			return nil, mocks.GenericError
 		}
 
@@ -1166,10 +1166,10 @@ func TestServer_ExecuteScriptAtLatestBlock(t *testing.T) {
 		t.Parallel()
 
 		invoker := mocks.BaselineInvoker(t)
-		invoker.ScriptFunc = func(height uint64, script []byte, parameters []cadence.Value) (cadence.Value, error) {
+		invoker.ScriptFunc = func(height uint64, script []byte, parameters [][]byte) (cadence.Value, error) {
 			assert.Equal(t, mocks.GenericHeight, height)
 			assert.Equal(t, mocks.GenericBytes, script)
-			assert.Equal(t, []cadence.Value{cadenceValue}, parameters)
+			assert.Equal(t, [][]byte{cadenceValueBytes}, parameters)
 
 			return mocks.GenericAmount(0), nil
 		}
@@ -1213,7 +1213,7 @@ func TestServer_ExecuteScriptAtLatestBlock(t *testing.T) {
 		t.Parallel()
 
 		invoker := mocks.BaselineInvoker(t)
-		invoker.ScriptFunc = func(uint64, []byte, []cadence.Value) (cadence.Value, error) {
+		invoker.ScriptFunc = func(uint64, []byte, [][]byte) (cadence.Value, error) {
 			return nil, mocks.GenericError
 		}
