+++ service/invoker/invoker.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/dgraph-io/ristretto"
 	"github.com/onflow/cadence"
-	"github.com/onflow/cadence/encoding/json"
 	"github.com/onflow/flow-archive/util"
 	"github.com/onflow/flow-go/fvm"
 	"github.com/onflow/flow-go/fvm/environment"
@@ -21,6 +20,7 @@ import (
 type Invoker struct {
 	index archive.Reader
 	vm    fvm.VM
+	vmCtx fvm.Context
 	cache Cache
 }
 
@@ -61,45 +61,40 @@ func New(index archive.Reader, options ...func(*Config)) (*Invoker, error) {
 		cache: cache,
 	}
 
+	vmCtx := fvm.NewContext(fvm.WithBlocks(&i))
+
+	i.vmCtx = vmCtx
+
 	return &i, nil
 }
 
 // Key returns the public key of the account with the given address.
 func (i *Invoker) Key(height uint64, address flow.Address, index int) (*flow.AccountPublicKey, error) {
-	err := util.ValidateHeightIndexed(i.index, height)
-	if err != nil {
-		return nil, fmt.Errorf("data unavailable for block height: %w", err)
-	}
-
 	// Retrieve the account at the specified block height.
 	account, err := i.Account(height, address)
 	if err != nil {
 		return nil, fmt.Errorf("could not retrieve account: %w", err)
 	}
 
-	// Create a key lookup map and check if the requested key exists.
-	keys := make(map[int]flow.AccountPublicKey)
 	for _, key := range account.Keys {
-		keys[key.Index] = key
-	}
-	key, ok := keys[index]
-	if !ok {
-		return nil, fmt.Errorf("account key with given index not found")
+		if key.Index != index {
+			continue
+		}
+
+		if key.Revoked {
+			return nil, fmt.Errorf("account key with given index has been revoked")
+		}
+		return &key, nil
 	}
 
-	// Check if the key is still valid.
-	if key.Revoked {
-		return nil, fmt.Errorf("account key with given index has been revoked")
-	}
-
-	return &key, nil
+	return nil, fmt.Errorf("account key with given index not found")
 }
 
 // Account returns the account with the given address.
 func (i *Invoker) Account(height uint64, address flow.Address) (*flow.Account, error) {
-	err := util.ValidateHeightIndexed(i.index, height)
+	err := util.ValidateHeightDataAvailable(i.index, height)
 	if err != nil {
-		return nil, fmt.Errorf("data unavailable for block height: %w", err)
+		return nil, err
 	}
 	// Look up the current block and commit for the block.
 	header, err := i.index.Header(height)
@@ -107,17 +102,9 @@ func (i *Invoker) Account(height uint64, address flow.Address) (*flow.Account, e
 		return nil, fmt.Errorf("could not get header: %w", err)
 	}
 
-	ctx := fvm.NewContext(fvm.WithBlockHeader(header), fvm.WithBlocks(i))
+	ctx := fvm.NewContextFromParent(i.vmCtx, fvm.WithBlockHeader(header))
 
-	// Initialize the storage snapshot. We use a shared cache between all
-	// heights here. It's a smart cache, which means that items that are
-	// accessed often are more likely to be kept, regardless of height. This
-	// allows us to put an upper bound on total cache size while using it for
-	// all heights.
-	storageSnapshot := snapshot.NewReadFuncStorageSnapshot(
-		readRegister(i.index, i.cache, header.Height))
-
-	account, err := i.vm.GetAccount(ctx, address, storageSnapshot)
+	account, err := i.vm.GetAccount(ctx, address, i.storageSnapshot(height))
 	if err != nil {
 		return nil, fmt.Errorf("could not get account at height %d: %w", header.Height, err)
 	}
@@ -126,20 +113,10 @@ func (i *Invoker) Account(height uint64, address flow.Address) (*flow.Account, e
 }
 
 // Script executes the given Cadence script and returns its result.
-func (i *Invoker) Script(height uint64, script []byte, arguments []cadence.Value) (cadence.Value, error) {
-
-	// Encode the arguments from Cadence values to byte slices.
-	var args [][]byte
-	for _, argument := range arguments {
-		arg, err := json.Encode(argument)
-		if err != nil {
-			return nil, fmt.Errorf("could not encode value: %w", err)
-		}
-		args = append(args, arg)
-	}
-	err := util.ValidateHeightIndexed(i.index, height)
+func (i *Invoker) Script(height uint64, script []byte, args [][]byte) (cadence.Value, error) {
+	err := util.ValidateHeightDataAvailable(i.index, height)
 	if err != nil {
-		return nil, fmt.Errorf("data unavailable for block height: %w", err)
+		return nil, err
 	}
 	// Look up the current block and commit for the block.
 	header, err := i.index.Header(height)
@@ -147,17 +124,7 @@ func (i *Invoker) Script(height uint64, script []byte, arguments []cadence.Value
 		return nil, fmt.Errorf("could not get header: %w", err)
 	}
 
-	// Initialize the virtual machine context with the given block header so
-	// that parameters related to the block are available from within the script.
-	ctx := fvm.NewContext(fvm.WithBlockHeader(header), fvm.WithBlocks(i))
-
-	// Initialize the storage snapshot. We use a shared cache between all
-	// heights here. It's a smart cache, which means that items that are
-	// accessed often are more likely to be kept, regardless of height. This
-	// allows us to put an upper bound on total cache size while using it for
-	// all heights.
-	storageSnapshot := snapshot.NewReadFuncStorageSnapshot(
-		readRegister(i.index, i.cache, height))
+	ctx := fvm.NewContextFromParent(i.vmCtx, fvm.WithBlockHeader(header))
 
 	// Initialize the procedure using the script bytes and the encoded
 	// Cadence parameters.
@@ -165,7 +132,7 @@ func (i *Invoker) Script(height uint64, script []byte, arguments []cadence.Value
 
 	// The script procedure is then run using the Flow virtual machine and all
 	// the constructed contextual parameters.
-	_, output, err := i.vm.Run(ctx, proc, storageSnapshot)
+	_, output, err := i.vm.Run(ctx, proc, i.storageSnapshot(height))
 	if err != nil {
 		return nil, fmt.Errorf("could not run script: %w", err)
 	}
@@ -194,4 +161,41 @@ func (i *Invoker) ByHeightFrom(height uint64, header *flow.Header) (*flow.Header
 	}
 	// find the block in storage as all of them are guaranteed to be finalized
 	return i.index.Header(height)
+}
+
+func (i *Invoker) storageSnapshot(height uint64) snapshot.StorageSnapshot {
+	// Initialize the storage snapshot. We use a shared cache between all
+	// heights here. It's a smart cache, which means that items that are
+	// accessed often are more likely to be kept, regardless of height. This
+	// allows us to put an upper bound on total cache size while using it for
+	// all heights.
+	return snapshot.NewReadFuncStorageSnapshot(
+		readRegister(i.index, i.cache, height))
+}
+
+func readRegister(
+	index archive.Reader,
+	cache Cache,
+	height uint64,
+) func(flow.RegisterID) (flow.RegisterValue, error) {
+	return func(regID flow.RegisterID) (flow.RegisterValue, error) {
+		cacheKey := fmt.Sprintf("%d/%s", height, regID)
+		cacheValue, ok := cache.Get(cacheKey)
+		if ok {
+			return cacheValue.(flow.RegisterValue), nil
+		}
+
+		values, err := index.Values(height, flow.RegisterIDs{regID})
+		if err != nil {
+			return nil, fmt.Errorf("could not read register: %w", err)
+		}
+		if len(values) != 1 {
+			return nil, fmt.Errorf("wrong number of register values: %d", len(values))
+		}
+
+		value := values[0]
+		_ = cache.Set(cacheKey, value, int64(len(value)))
+
+		return value, nil
+	}
 }
